WIP - Bug 573766: Terminate All: consider number of non-terminated launches only
The actual selection is irrelevant to whether or not the action should
be enabled - i.e. right-clicking in white space in Debug view, or
using the Terminate/Disconnect All command directly (via Ctrl-3) should
terminate everything.

This commit is WIP because the current implementation still requires
that a Debug View (aka LaunchView) is accessible on the current page - see
org.eclipse.debug.internal.ui.commands.actions.DebugActionHandler.getDelegate()
which gets the TerminateAllAction from the Debug View. I don't understand
why the Debug View is the keeper of these objects as the actual actions
could just be globally maintained. In the specific case of Terminate All,
which should be always available, the action is not even created in
the first place if the Debug View has never been shown*.

* You can launch a program, if it doesn't hit a breakpoint, the Debub View
won't open and the terminate all action won't be created which means
that the there is no way to get that action.

All of this leads me to the conclusion that the TerminateAll action
should not actually be a DebugCommandAction at all and actually be
more like the AbstractRemoveAllActionDelegate which exists independently
of the debug view.

The first part of that is TerminateAllActionDelegate2 - and if that looks
to be on the correct path then the TerminateAllAction can be removed.

Change-Id: I56505943195aa2bb74a5703696d7ba62d08d208b